requirements.txt
fastai>=2.7.0
bing-image-downloader
Pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.5.0